requirements.txt
Flask
speedtest-cli==2.1.3
bleak
